+++ src/main.tsx
@@ -3,6 +3,9 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
+import { initTelemetry, readTelemetryConsent } from '@/telemetry'
+
+void initTelemetry(readTelemetryConsent())
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>